Add in shipping page changes to improve user workflow and allow deletion if user chooses to deliver something else, fixed issues with qty counter in shipping page and add in extra functionalities

+++ json/inventory-delete.php
@@ -24,9 +24,10 @@ $rootdir = $_SERVER['ROOT_DIR'];
 	$partid = grab('partid');
 	$serial = grab('serial');
 	$po_number = grab('po_number');
+	$page = grab('page');
 	
 	//items = ['partid', 'serial', 'qty', 'location', 'status', 'condition'];
-	function deletefromDatabase($partid, $serial, $po_number){
+	function deletefromDatabase($partid, $serial, $po_number, $page) {
 		$result = array();
 		
 		$result['query'] = true;
@@ -38,14 +39,27 @@ $rootdir = $_SERVER['ROOT_DIR'];
 		
 		//If the item exists then delete the inventory row
 		if($check->num_rows == 1) {
-			//Delete the inventory item
-			$query = "DELETE FROM inventory WHERE partid = '" . res($partid) . "' AND serial_no = '" . res($serial) . "';";
-			$result['query'] = qdb($query);
-
-			//If the item is deleted from the inventory then increment the purchase items back to original state, before the serial was scanned
-			if($result['query']) {
-				$query = "UPDATE purchase_items SET qty_received = qty_received - 1 WHERE po_number = ". res($po_number) ." AND partid = ". res($partid) .";";
-				qdb($query);
+			//Check if the page is shipping or not (Purchase Order or Sales Order receive/ship)
+			if($page != 'shipping') {
+				//Delete the inventory item
+				$query = "DELETE FROM inventory WHERE partid = '" . res($partid) . "' AND serial_no = '" . res($serial) . "';";
+				$result['query'] = qdb($query);
+	
+				//If the item is deleted from the inventory then increment the purchase items back to original state, before the serial was scanned
+				if($result['query']) {
+					$query = "UPDATE purchase_items SET qty_received = qty_received - 1 WHERE po_number = ". res($po_number) ." AND partid = ". res($partid) .";";
+					qdb($query);
+				}
+			//Current process is shipping
+			} else {
+				$query = "UPDATE inventory SET qty = qty + 1, last_sale = NULL WHERE partid = '" . res($partid) . "' AND serial_no = '" . res($serial) . "';";
+				$result['query'] = qdb($query);
+	
+				//If the item is deleted from the inventory then increment the purchase items back to original state, before the serial was scanned
+				if($result['query']) {
+					$query = "UPDATE sales_items SET qty_shipped = qty_shipped - 1, ship_date = NULL WHERE so_number = ". res($po_number) ." AND partid = ". res($partid) .";";
+					qdb($query);
+				}
 			}
 		} else {
 			$result['query'] = false;
@@ -54,6 +68,6 @@ $rootdir = $_SERVER['ROOT_DIR'];
 		return $result;
 	}
 	
-	$result = deletefromDatabase($partid, $serial, $po_number);
+	$result = deletefromDatabase($partid, $serial, $po_number, $page);
 	echo json_encode($result);
     exit;
